M2-6217: cleaned up cycles in utils
