+++ src/shared/components/NoPermissionPopup/NoPermissionPopup.tsx
@@ -1,7 +1,7 @@
 import { useTranslation } from 'react-i18next';
 
 import { falseReturnFunc } from 'shared/utils';
-import { Modal } from 'shared/components';
+import { Modal } from 'shared/components/Modal';
 import { StyledModalWrapper } from 'shared/styles';
 
 import { useNoPermissionPopup } from './NoPermissionPopup.hooks';
